Add support for searching the address book
